make some updates to the README
I just cloned this repo for the first time in a while on a new computer,
and realized that the README wasn't accurate. In particular, it tells
the user to set a `-d` flag to tell `ocw-to-hugo` to download courses,
but the flag is actually `--download`. So I updated the README to
reflect that, and did a little bit of formatting too.

pr #419